add base sampler class
