+++ evalml/pipelines/components/transformers/samplers/undersampler.py
@@ -1,17 +1,13 @@
 import pandas as pd
 from skopt.space import Integer, Real
 
-from evalml.pipelines.components.transformers import Transformer
+from evalml.pipelines.components.transformers.samplers.base_sampler import BaseSampler
 from evalml.preprocessing.data_splitters.balanced_classification_sampler import (
     BalancedClassificationSampler
 )
-from evalml.utils.woodwork_utils import (
-    _convert_woodwork_types_wrapper,
-    infer_feature_types
-)
 
 
-class Undersampler(Transformer):
+class Undersampler(BaseSampler):
     """Random undersampler component."""
     name = "Undersampler"
     hyperparameter_ranges = {
@@ -43,20 +39,6 @@ class Undersampler(Transformer):
                          component_obj=sampler,
                          random_seed=random_seed)
 
-    def fit(self, X, y):
-        """Resample the data using the undersampler. Since our sampler doesn't need to be fit, we do nothing here.
-
-        Arguments:
-            X (ww.DataFrame): Training features
-            y (ww.DataColumn): Target features
-
-        Returns:
-            self
-        """
-        if y is None:
-            raise ValueError("y cannot be none")
-        return self
-
     def fit_transform(self, X, y):
         """Fit and transform the data using the undersampler. Used during training of the pipeline
 
@@ -67,28 +49,8 @@ class Undersampler(Transformer):
          Returns:
             ww.DataTable, ww.DataColumn: Undersampled X and y data
         """
-        X = infer_feature_types(X)
-        if y is None:
-            raise ValueError("y cannot be none")
-        y = infer_feature_types(y)
-        X_pd = _convert_woodwork_types_wrapper(X.to_dataframe())
-        y_pd = _convert_woodwork_types_wrapper(y.to_series())
+        X_pd, y_pd = self._prepare_data(X, y)
         index_df = pd.Series(y_pd.index)
         indices = self._component_obj.fit_resample(X_pd, y_pd)
         train_indices = index_df[index_df.isin(indices)].index.values.tolist()
         return X.iloc[train_indices], y.iloc[train_indices]
-
-    def transform(self, X, y=None):
-        """No transformation needs to be done here.
-
-        Arguments:
-            X (ww.DataFrame): Training features. Ignored.
-            y (ww.DataColumn): Target features. Ignored.
-
-        Returns:
-            ww.DataTable, ww.DataColumn: X and y data that was passed in.
-        """
-        X = infer_feature_types(X)
-        if y is not None:
-            y = infer_feature_types(y)
-        return X, y
